Add ability to pass custom domains
Committing this in its own branch but I won’t be merging this into master as I don’t see a use case for it at the moment given that if used with a service like ngrok, it results in certificate warnings. The way have publicly-accessible staging is to use ngrok and integrate with automatic Let’s Encrypt certificate generation using Greenlock.js.